+++ core/libraries/bancha/dispatchers/dispatcher_default.php
@@ -21,8 +21,11 @@ Class Dispatcher_default
 
 	public function __construct()
 	{
-		$this->_CI = & get_instance();
+		$this->_CI = & CI_Controller::get_instance();
 		$this->_CI->load->database();
+
+		//Front-end helper
+		$this->_CI->load->helper('frontend');
 	}
 
 	/**
@@ -106,6 +109,8 @@ Class Dispatcher_default
 				$this->_CI->load->dispatcher('print', 'dispatcher_print');
 				$this->_CI->dispatcher_print->render($page);
 			} else {
+				//We add a page global variable
+				$GLOBALS['page'] = & $page;
 				$this->_CI->view->set('page', $page);
 				$this->_CI->view->render_template($page->get('view_template'));
 			}
@@ -149,6 +154,7 @@ Class Dispatcher_default
 
 		$parent_page->set('view_template', $template);
 		$this->_CI->view->set('record', $record);
+		$GLOBALS['record'] = & $record;
 
 		return $parent_page;
 	}
@@ -310,8 +316,11 @@ Class Dispatcher_default
 
 			if ($this->_CI->config->item('type_custom_feeds') && isset($type['name']))
 			{
+				$page->set('records', $records);
 				$this->_CI->view->set('page', $page);
+				$GLOBALS['page'] = & $page;
 				$this->_CI->view->set('records', $records);
+				$GLOBALS['records'] = & $records;
 				$this->_CI->view->render_type_template($type['name'], 'feed', TRUE);
 				return;
 			}