+++ themes/sandbox/views/footer.php
@@ -1,0 +1,4 @@
+<footer>
+	<hr />
+	<p>Bancha - Generated in {elapsed_time} seconds. - <?php languages(); ?></p>
+</footer>