+++ themes/sandbox/views/type_templates/Default-Page/detail.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * Detail View
+ *
+ * Page detail - Boilerplate Theme
+ *
+ * @package		Bancha
+ * @author		Nicholas Valbusa - info@squallstar.it - @squallstar
+ * @copyright	Copyright (c) 2011, Squallstar
+ * @license		GNU/GPL (General Public License)
+ *
+ */
+
+?>
+
+<h1><?php echo page('title'); ?></h1>
+
+<p class="submenu">
+	<?php echo menu(tree('current')); ?>
+</p>
+
+<p><?php echo page('content'); ?></p>
+
+<?php
+/* End of file detail.php */
+/* Location: /type_templates/Default-Page/detail.php */